+++ packages/flutter_adaptive_scaffold/lib/src/breakpoints.dart
@@ -4,6 +4,8 @@
 
 import 'package:flutter/material.dart';
 
+import '../flutter_adaptive_scaffold.dart';
+
 /// A group of standard breakpoints built according to the material
 /// specifications for screen width size.
 ///
@@ -15,7 +17,7 @@ class Breakpoints {
   /// case that no other breakpoint is active.
   ///
   /// It is active from a width of -1 dp to infinity.
-  static const Breakpoint standard = Breakpoint(beginWidth: -1);
+  static const Breakpoint standard = Breakpoint.standard();
 
   /// A window whose width is less than 600 dp and greater than 0 dp.
   static const Breakpoint small = Breakpoint.small();
@@ -84,6 +86,30 @@ class Breakpoints {
   /// A mobile window whose width is greater than 1600 dp.
   static const Breakpoint extraLargeMobile =
       Breakpoint.extraLarge(platform: Breakpoint.mobile);
+
+  /// A list of all the standard breakpoints.
+  static const List<Breakpoint> all = <Breakpoint>[
+    smallDesktop,
+    smallMobile,
+    small,
+    mediumDesktop,
+    mediumMobile,
+    medium,
+    mediumLargeDesktop,
+    mediumLargeMobile,
+    mediumLarge,
+    largeDesktop,
+    largeMobile,
+    large,
+    extraLargeDesktop,
+    extraLargeMobile,
+    extraLarge,
+    smallAndUp,
+    mediumAndUp,
+    mediumLargeAndUp,
+    largeAndUp,
+    standard,
+  ];
 }
 
 /// A class to define the conditions that distinguish between types of
@@ -111,9 +137,18 @@ class Breakpoint {
     this.endWidth,
     this.beginHeight,
     this.endHeight,
-    this.platform,
     this.andUp = false,
+    this.platform,
   });
+
+  /// Returns a [Breakpoint] that can be used as a fallthrough in the
+  /// case that no other breakpoint is active.
+  const Breakpoint.standard({this.platform})
+      : beginWidth = -1,
+        endWidth = null,
+        beginHeight = null,
+        endHeight = null,
+        andUp = true;
 
   /// Returns a [Breakpoint] with the given constraints for a small screen.
   const Breakpoint.small({this.andUp = false, this.platform})
@@ -164,7 +199,7 @@ class Breakpoint {
     TargetPlatform.iOS,
   };
 
-  /// When set to true, it will include any size above the set width.
+  /// When set to true, it will include any size above the set width and set height.
   final bool andUp;
 
   /// The beginning width dp value. If left null then the [Breakpoint] will have
@@ -193,6 +228,7 @@ class Breakpoint {
   bool isActive(BuildContext context) {
     final TargetPlatform host = Theme.of(context).platform;
     final bool isRightPlatform = platform?.contains(host) ?? true;
+    final bool isDesktop = Breakpoint.desktop.contains(host);
 
     final double width = MediaQuery.sizeOf(context).width;
     final double height = MediaQuery.sizeOf(context).height;
@@ -208,11 +244,54 @@ class Breakpoint {
         ? width >= lowerBoundWidth
         : width >= lowerBoundWidth && width < upperBoundWidth;
 
-    final bool isHeightActive = (orientation == Orientation.landscape &&
-            height >= lowerBoundHeight &&
-            height < upperBoundHeight) ||
-        orientation == Orientation.portrait;
+    final bool isHeightActive = isDesktop ||
+        orientation == Orientation.portrait ||
+        (orientation == Orientation.landscape && andUp
+            ? isWidthActive || height >= lowerBoundHeight
+            : height >= lowerBoundHeight && height < upperBoundHeight);
 
     return isWidthActive && isHeightActive && isRightPlatform;
+  }
+
+  /// Returns the currently active [Breakpoint] based on the [SlotLayout] in the
+  /// context.
+  static Breakpoint? maybeActiveBreakpointFromSlotLayout(BuildContext context) {
+    final SlotLayout? slotLayout =
+        context.findAncestorWidgetOfExactType<SlotLayout>();
+
+    return slotLayout != null
+        ? activeBreakpointIn(context, slotLayout.config.keys.toList())
+        : null;
+  }
+
+  /// Returns the default [Breakpoint] based on the [BuildContext].
+  static Breakpoint defaultBreakpointOf(BuildContext context) {
+    return activeBreakpointIn(context, Breakpoints.all) ?? Breakpoints.standard;
+  }
+
+  /// Returns the currently active [Breakpoint].
+  static Breakpoint activeBreakpointOf(BuildContext context) {
+    return maybeActiveBreakpointFromSlotLayout(context) ??
+        defaultBreakpointOf(context);
+  }
+
+  /// Returns the currently active [Breakpoint] based on the [BuildContext] and
+  /// a list of [Breakpoint]s.
+  static Breakpoint? activeBreakpointIn(
+      BuildContext context, List<Breakpoint> breakpoints) {
+    Breakpoint? currentBreakpoint;
+
+    for (final Breakpoint breakpoint in breakpoints) {
+      if (breakpoint.isActive(context)) {
+        if (breakpoint.platform != null) {
+          // Prioritize platform-specific breakpoints.
+          return breakpoint;
+        } else {
+          // Fallback to non-platform-specific.
+          currentBreakpoint = breakpoint;
+        }
+      }
+    }
+    return currentBreakpoint;
   }
 }